MAINT: group tests of reductions in tests/test_reductions.py
