+++ torch_np/tests/test_ndarray_methods.py
@@ -518,78 +518,3 @@ class TestArgmin:
         a = np.array([1, -2**63, -2**63 + 1, 2**63 - 1], dtype=np.int64)
         assert_equal(np.argmin(a), 1)
 
-
-class TestAny:
-    def test_basic(self):
-        y1 = [0, 0, 1, 0]
-        y2 = [0, 0, 0, 0]
-        y3 = [1, 0, 1, 0]
-        assert np.any(y1)
-        assert np.any(y3)
-        assert not np.any(y2)
-
-    def test_nd(self):
-        y1 = [[0, 0, 0], [0, 1, 0], [1, 1, 0]]
-        assert np.any(y1)
-        assert_equal(np.any(y1, axis=0), [1, 1, 0])
-        assert_equal(np.any(y1, axis=1), [0, 1, 1])
-        assert_equal(np.any(y1), True)
-        assert isinstance(np.any(y1, axis=1), np.ndarray)
-
-    # YYY: deduplicate 
-    def test_method_vs_function(self):
-        y = np.array([[0, 1, 0, 3], [1, 0, 2, 0]])
-        assert_equal(np.any(y), y.any())
-
-    def test_any_bad_axis(self):
-        # Basic check of functionality
-        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
-
-        assert_raises(TypeError, np.any, m, axis='foo')
-        assert_raises(np.AxisError, np.any, m, axis=3)
-        assert_raises(TypeError, np.any,
-                      m, axis=np.array([[1], [2]]))
-
-    @pytest.mark.xfail(reason='XXX: pytorch does not support any(..., axis=tuple)')
-    def test_any_axis_bad_tuple(self):
-        # Basic check of functionality
-        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
-        assert_raises(ValueError, np.any, m, axis=(1, 1))
-
-
-class TestAll:
-    def test_basic(self):
-        y1 = [0, 1, 1, 0]
-        y2 = [0, 0, 0, 0]
-        y3 = [1, 1, 1, 1]
-        assert not np.all(y1)
-        assert np.all(y3)
-        assert not np.all(y2)
-        assert np.all(~np.array(y2))
-
-    def test_nd(self):
-        y1 = [[0, 0, 1], [0, 1, 1], [1, 1, 1]]
-        assert not np.all(y1)
-        assert_equal(np.all(y1, axis=0), [0, 0, 1])
-        assert_equal(np.all(y1, axis=1), [0, 0, 1])
-        assert_equal(np.all(y1), False)
-
-    def test_method_vs_function(self):
-        y = np.array([[0, 1, 0, 3], [1, 0, 2, 0]])
-        assert_equal(np.all(y), y.all())
-
-    def test_any_bad_axis(self):
-        # Basic check of functionality
-        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
-
-        assert_raises(TypeError, np.all, m, axis='foo')
-        assert_raises(np.AxisError, np.all, m, axis=3)
-        assert_raises(TypeError, np.all,
-                      m, axis=np.array([[1], [2]]))
-
-    @pytest.mark.xfail(reason='XXX: pytorch does not support all(..., axis=tuple)')
-    def test_any_axis_bad_tuple(self):
-        # Basic check of functionality
-        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
-        assert_raises(ValueError, np.all, m, axis=(1, 1))
-
