+++ torch_np/tests/test_reductions.py
@@ -1,0 +1,269 @@
+import pytest
+from pytest import raises as assert_raises
+
+import torch_np as np
+from torch_np.testing import assert_equal, assert_array_equal
+
+
+class TestNonzeroAndCountNonzero:
+    def test_nonzero_trivial(self):
+        assert_equal(np.count_nonzero(np.array([])), 0)
+        assert_equal(np.count_nonzero(np.array([], dtype='?')), 0)
+        assert_equal(np.nonzero(np.array([])), ([],))
+
+        assert_equal(np.count_nonzero(np.array([0])), 0)
+        assert_equal(np.count_nonzero(np.array([0], dtype='?')), 0)
+        assert_equal(np.nonzero(np.array([0])), ([],))
+
+        assert_equal(np.count_nonzero(np.array([1])), 1)
+        assert_equal(np.count_nonzero(np.array([1], dtype='?')), 1)
+        assert_equal(np.nonzero(np.array([1])), ([0],))
+
+        assert isinstance(np.count_nonzero([]), np.ndarray)
+
+    def test_nonzero_zerod(self):
+        assert_equal(np.count_nonzero(np.array(0)), 0)
+        assert_equal(np.count_nonzero(np.array(0, dtype='?')), 0)
+
+        assert_equal(np.count_nonzero(np.array(1)), 1)
+        assert_equal(np.count_nonzero(np.array(1, dtype='?')), 1)
+
+        assert isinstance(np.count_nonzero(np.array(1)), np.ndarray)
+
+    def test_nonzero_onedim(self):
+        x = np.array([1, 0, 2, -1, 0, 0, 8])
+        assert_equal(np.count_nonzero(x), 4)
+        assert_equal(np.count_nonzero(x), 4)
+        assert_equal(np.nonzero(x), ([0, 2, 3, 6],))
+
+        assert isinstance(np.count_nonzero(x), np.ndarray)
+
+    def test_nonzero_twodim(self):
+        x = np.array([[0, 1, 0], [2, 0, 3]])
+        assert_equal(np.count_nonzero(x.astype('i1')), 3)
+        assert_equal(np.count_nonzero(x.astype('i2')), 3)
+        assert_equal(np.count_nonzero(x.astype('i4')), 3)
+        assert_equal(np.count_nonzero(x.astype('i8')), 3)
+        assert_equal(np.nonzero(x), ([0, 1, 1], [1, 0, 2]))
+
+        x = np.eye(3)
+        assert_equal(np.count_nonzero(x.astype('i1')), 3)
+        assert_equal(np.count_nonzero(x.astype('i2')), 3)
+        assert_equal(np.count_nonzero(x.astype('i4')), 3)
+        assert_equal(np.count_nonzero(x.astype('i8')), 3)
+        assert_equal(np.nonzero(x), ([0, 1, 2], [0, 1, 2]))
+
+    def test_sparse(self):
+        # test special sparse condition boolean code path
+        for i in range(20):
+            c = np.zeros(200, dtype=bool)
+            c[i::20] = True
+            assert_equal(np.nonzero(c)[0], np.arange(i, 200 + i, 20))
+
+            c = np.zeros(400, dtype=bool)
+            c[10 + i:20 + i] = True
+            c[20 + i*2] = True
+            assert_equal(np.nonzero(c)[0],
+                         np.concatenate((np.arange(10 + i, 20 + i), [20 + i*2])))
+
+    def test_count_nonzero_axis(self):
+        # Basic check of functionality
+        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
+
+        expected = np.array([1, 1, 1, 1, 1])
+        assert_equal(np.count_nonzero(m, axis=0), expected)
+
+        expected = np.array([2, 3])
+        assert_equal(np.count_nonzero(m, axis=1), expected)
+
+        assert isinstance(np.count_nonzero(m, axis=1), np.ndarray)
+
+        assert_raises(ValueError, np.count_nonzero, m, axis=(1, 1))
+        assert_raises(TypeError, np.count_nonzero, m, axis='foo')
+        assert_raises(np.AxisError, np.count_nonzero, m, axis=3)
+        assert_raises(TypeError, np.count_nonzero,
+                      m, axis=np.array([[1], [2]]))
+
+    @pytest.mark.parametrize('typecode', np.typecodes['All'])
+    def test_count_nonzero_axis_all_dtypes(self, typecode):
+        # More thorough test that the axis argument is respected
+        # for all dtypes and responds correctly when presented with
+        # either integer or tuple arguments for axis
+
+        m = np.zeros((3, 3), dtype=typecode)
+        n = np.ones(1, dtype=typecode)
+
+        m[0, 0] = n[0]
+        m[1, 0] = n[0]
+
+        expected = np.array([2, 0, 0], dtype=np.intp)
+        result = np.count_nonzero(m, axis=0)
+        assert_equal(result, expected)
+        assert expected.dtype == result.dtype
+
+        expected = np.array([1, 1, 0], dtype=np.intp)
+        result = np.count_nonzero(m, axis=1)
+        assert_equal(result, expected)
+        assert expected.dtype == result.dtype
+
+        expected = np.array(2)
+        assert_equal(np.count_nonzero(m, axis=(0, 1)),
+                     expected)
+        assert_equal(np.count_nonzero(m, axis=None),
+                     expected)
+        assert_equal(np.count_nonzero(m),
+                     expected)
+
+    def test_countnonzero_axis_empty(self):
+        a = np.array([[0, 0, 1], [1, 0, 1]])
+        assert_equal(np.count_nonzero(a, axis=()), a.astype(bool))
+
+    def test_countnonzero_keepdims(self):
+        a = np.array([[0, 0, 1, 0],
+                      [0, 3, 5, 0],
+                      [7, 9, 2, 0]])
+        assert_equal(np.count_nonzero(a, axis=0, keepdims=True),
+                     [[1, 2, 3, 0]])
+        assert_equal(np.count_nonzero(a, axis=1, keepdims=True),
+                     [[1], [2], [3]])
+        assert_equal(np.count_nonzero(a, keepdims=True),
+                     [[6]])
+        assert isinstance(np.count_nonzero(a, axis=1, keepdims=True), np.ndarray)
+
+
+    @pytest.mark.parametrize('axis',
+            [0, 1, 2, -1, -2, None, (), (0, 1), (1, 0), (0, 1, 2), (1, -1, 0)])
+    def count_nonzero_keepdims2(self, axis):
+        a = np.arange(2*3*4).reshape((2, 3, 4))
+        with_keepdims = np.count_nonzero(a, axis, keepdims=True)
+        expanded = np.expand_dims( np.count_nonzero(a, axis=axis), axis=axis)
+        assert_equal(with_keepdims, expanded)
+
+    def test_array_method(self):
+        # Tests that the array method
+        # call to nonzero works
+        m = np.array([[1, 0, 0], [4, 0, 6]])
+        tgt = [[0, 1, 1], [0, 0, 2]]
+
+        assert_equal(m.nonzero(), tgt)
+
+
+class TestFlatnonzero:
+    def test_basic(self):
+        x = np.arange(-2, 3)
+        assert_equal(np.flatnonzero(x),
+                     [0, 1, 3, 4])
+
+
+class TestArgwhere:
+
+    @pytest.mark.parametrize('nd', [0, 1, 2])
+    def test_nd(self, nd):
+        # get an nd array with multiple elements in every dimension
+        x = np.empty((2,)*nd, bool)
+
+        # none
+        x[...] = False
+        assert_equal(np.argwhere(x).shape, (0, nd))
+
+        # only one
+        x[...] = False
+        x.ravel()[0] = True
+        assert_equal(np.argwhere(x).shape, (1, nd))
+
+        # all but one
+        x[...] = True
+        x.ravel()[0] = False
+        assert_equal(np.argwhere(x).shape, (x.size - 1, nd))
+
+        # all
+        x[...] = True
+        assert_equal(np.argwhere(x).shape, (x.size, nd))
+
+    def test_2D(self):
+        x = np.arange(6).reshape((2, 3))
+        assert_array_equal(np.argwhere(x > 1),
+                           [[0, 2],
+                            [1, 0],
+                            [1, 1],
+                            [1, 2]])
+
+    def test_list(self):
+        assert_equal(np.argwhere([4, 0, 2, 1, 3]), [[0], [2], [3], [4]])
+
+
+class TestAny:
+    def test_basic(self):
+        y1 = [0, 0, 1, 0]
+        y2 = [0, 0, 0, 0]
+        y3 = [1, 0, 1, 0]
+        assert np.any(y1)
+        assert np.any(y3)
+        assert not np.any(y2)
+
+    def test_nd(self):
+        y1 = [[0, 0, 0], [0, 1, 0], [1, 1, 0]]
+        assert np.any(y1)
+        assert_equal(np.any(y1, axis=0), [1, 1, 0])
+        assert_equal(np.any(y1, axis=1), [0, 1, 1])
+        assert_equal(np.any(y1), True)
+        assert isinstance(np.any(y1, axis=1), np.ndarray)
+
+    # YYY: deduplicate 
+    def test_method_vs_function(self):
+        y = np.array([[0, 1, 0, 3], [1, 0, 2, 0]])
+        assert_equal(np.any(y), y.any())
+
+    def test_any_bad_axis(self):
+        # Basic check of functionality
+        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
+
+        assert_raises(TypeError, np.any, m, axis='foo')
+        assert_raises(np.AxisError, np.any, m, axis=3)
+        assert_raises(TypeError, np.any,
+                      m, axis=np.array([[1], [2]]))
+
+    @pytest.mark.xfail(reason='XXX: pytorch does not support any(..., axis=tuple)')
+    def test_any_axis_bad_tuple(self):
+        # Basic check of functionality
+        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
+        assert_raises(ValueError, np.any, m, axis=(1, 1))
+
+
+class TestAll:
+    def test_basic(self):
+        y1 = [0, 1, 1, 0]
+        y2 = [0, 0, 0, 0]
+        y3 = [1, 1, 1, 1]
+        assert not np.all(y1)
+        assert np.all(y3)
+        assert not np.all(y2)
+        assert np.all(~np.array(y2))
+
+    def test_nd(self):
+        y1 = [[0, 0, 1], [0, 1, 1], [1, 1, 1]]
+        assert not np.all(y1)
+        assert_equal(np.all(y1, axis=0), [0, 0, 1])
+        assert_equal(np.all(y1, axis=1), [0, 0, 1])
+        assert_equal(np.all(y1), False)
+
+    def test_method_vs_function(self):
+        y = np.array([[0, 1, 0, 3], [1, 0, 2, 0]])
+        assert_equal(np.all(y), y.all())
+
+    def test_any_bad_axis(self):
+        # Basic check of functionality
+        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
+
+        assert_raises(TypeError, np.all, m, axis='foo')
+        assert_raises(np.AxisError, np.all, m, axis=3)
+        assert_raises(TypeError, np.all,
+                      m, axis=np.array([[1], [2]]))
+
+    @pytest.mark.xfail(reason='XXX: pytorch does not support all(..., axis=tuple)')
+    def test_any_axis_bad_tuple(self):
+        # Basic check of functionality
+        m = np.array([[0, 1, 7, 0, 0], [3, 0, 0, 2, 19]])
+        assert_raises(ValueError, np.all, m, axis=(1, 1))
+
+
